feat: adding tests updating tests
feat: appsettings ++ UploadFolder configuration

chore: Update MaxFileSize constant to use long data type

fix: Remove unnecessary null assignment in ReceiptService

chore: Remove unnecessary code repetition in SchoolController

fix: Throw FileNotFoundException in FilePath.GetDataAsync if file does not exist

feat: Add GetDataAsStreamAsync method to FilePath model

+++ OgrenciAidatSistemi/Configurations/Constants.cs
@@ -34,6 +34,6 @@ namespace OgrenciAidatSistemi.Configurations
         public const String AuthenticationAccessDeniedPath = "/Error/Error403";
 
         // 10 MB file size
-        public const int MaxFileSize = 10 * 1024 * 1024;
+        public const long MaxFileSize = 10 * 1024 * 1024;
     }
 }
